Avoid circular dependency when defining MonadParse
