+++ server/src-lib/Hasura/GraphQL/Parser/Class.hs
@@ -1,5 +1,10 @@
 -- | Classes for monads used during schema construction and query parsing.
-module Hasura.GraphQL.Parser.Class where
+module Hasura.GraphQL.Parser.Class
+  ( MonadParse(..)
+  , parseError
+  , QueryReusability(..)
+  , module Hasura.GraphQL.Parser.Class
+  ) where
 
 import                          Hasura.Prelude
 
@@ -8,14 +13,14 @@ import                qualified Language.Haskell.TH                   as TH
 import                qualified Language.GraphQL.Draft.Syntax         as G
 
 import                          Data.Has
-import                          Data.Parser.JSONPath
 import                          Data.Text.Extended
 import                          Data.Tuple.Extended
 import                          GHC.Stack                             (HasCallStack)
 import                          Type.Reflection                       (Typeable)
 
 import                          Hasura.Backends.Postgres.SQL.Types
-import {-# SOURCE #-}           Hasura.GraphQL.Parser.Internal.Parser
+import                          Hasura.GraphQL.Parser.Class.Parse
+import                          Hasura.GraphQL.Parser.Internal.Types
 import                          Hasura.RQL.Types.Error
 import                          Hasura.RQL.Types.Table
 import                          Hasura.SQL.Backend
@@ -173,41 +178,3 @@ memoize4
   -> (a -> b -> c -> d -> m (Parser k n e))
   -> (a -> b -> c -> d -> m (Parser k n e))
 memoize4 name = curry4 . memoize name . uncurry4
-
--- | A class that provides functionality for parsing GraphQL queries, i.e.
--- running a fully-constructed 'Parser'.
-class Monad m => MonadParse m where
-  withPath :: (JSONPath -> JSONPath) -> m a -> m a
-  -- | Not the full power of 'MonadError' because parse errors cannot be
-  -- caught.
-  parseErrorWith :: Code -> Text -> m a
-  -- | See 'QueryReusability'.
-  markNotReusable :: m ()
-
-parseError :: MonadParse m => Text -> m a
-parseError = parseErrorWith ValidationFailed
-
--- | Tracks whether or not a query is /reusable/. Reusable queries are nice,
--- since we can cache their resolved ASTs and avoid re-resolving them if we
--- receive an identical query. However, we can’t always safely reuse queries if
--- they have variables, since some variable values can affect the generated SQL.
--- For example, consider the following query:
---
--- > query users_where($condition: users_bool_exp!) {
--- >   users(where: $condition) {
--- >     id
--- >   }
--- > }
---
--- Different values for @$condition@ will produce completely different queries,
--- so we can’t reuse its plan (unless the variable values were also all
--- identical, of course, but we don’t bother caching those).
-data QueryReusability = Reusable | NotReusable
-
-instance Semigroup QueryReusability where
-  NotReusable <> _           = NotReusable
-  _           <> NotReusable = NotReusable
-  Reusable    <> Reusable    = Reusable
-
-instance Monoid QueryReusability where
-  mempty = Reusable
