--- server/src-lib/Hasura/GraphQL/Parser/Class.hs-boot
@@ -1,5 +1,0 @@
-module Hasura.GraphQL.Parser.Class where
-
-import Data.Kind (Type)
-
-class MonadParse (m :: Type -> Type)
